SLP/requirements.txt
streamlit
pandas
supabase
google-api-python-client
pytrends
python-dotenv
